Conditionally test 2i index query
If the user is running the default riak backend (bitcask) or is testing
against a bucket with strong consistency, they will be able to send
index information with objects but it will be ignored. Modified the test
to check for the KRIA_TEST_2I env var. If it is present, the test
makes assertions about the result. If not, the test asserts that the
query returns no results.